Got queue to sync between browser sessions
So I did a bit of work on the queue, it's kind of messy but it works. I
did a little bit of formalization for the queue object so it's not just
a free-standing array, but not that much.

Now that the queue syncs though, if you'd like to do some testing
yourself or if your queue gets too cluttered, then type
chrome.storage.sync.clear() into the console and it will clear your
storage.