big changes to queue page
Okay so I got the queue to auto play and to play continuously through
the queued videos.
I took away the alert for youtube videos when you add them to queue so
it just adds it to queue.
The queue tab now only opens when you successfully add a link to the
queue.
If the queue is empty when you add your first item, it will begin auto
playing that item. If there are already items in the queue, then it will
begin playing at the last video you were at. When that video completes,
it will move on to the next video until the queue is exhausted.
You can add videos to the queue while the queue tab is playing and those
videos will still get played.